// src/lib/shopee/flash-sale/types.ts
export interface TimeSlot {
  timeslot_id: number;
  start_time: number;            // Unix timestamp
  end_time: number;              // Unix timestamp
}